+++ GameEngine/GameEngine/Engine/Utils/Serialization/SerializationMacros.hpp
@@ -163,3 +163,5 @@ class BinaryInputArchive; \
 			    }
 
 #define MACRO_STR(str)(#str)
+
+#define SERIALIZATION_CODE_BLOCK(code)(*"*/ code /*"*);
