+++ GameEngine/GameEngine/GameEngine.vcxproj.filters
@@ -91,6 +91,8 @@
     <ClCompile Include="Engine\Utils\Serialization\MatrixSerialization.cpp" />
     <ClCompile Include="Engine\Core\Tasks\RenderTasks.cpp" />
     <ClCompile Include="Engine\Geometry\Mesh.cpp" />
+    <ClCompile Include="Engine\Systems\FPControllerSystem.cpp" />
+    <ClCompile Include="Engine\Geometry\Material.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\Vendors\imgui\imconfig.h" />
